Move tests to folder
