+++ code/web/src/modules/user/tests/StylePreferences.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import StylePreferences from "./StylePreferences";
+import StylePreferences from "../StylePreferences";
 import { screen, render, fireEvent } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import { Provider } from "react-redux";
